requirements.txt
flask
opencv-python
numpy
joblib
scikit-learn
RPi.GPIO
pigpio
hx711